fix(toc): adds missing css
