--- packages/kit-headless/src/components/toc/index.ts
@@ -1,1 +1,0 @@
-export { TableOfContent } from './toc';
